+++ LinearMouse/UI/ButtonsSettings/ButtonMappingsSection/ButtonMappingAction/ButtonMappingAction.swift
@@ -1,0 +1,75 @@
+// MIT License
+// Copyright (c) 2021-2023 LinearMouse
+
+import SwiftUI
+
+struct ButtonMappingAction: View {
+    @Binding var action: Scheme.Buttons.Mapping.Action
+
+    var body: some View {
+        ButtonMappingActionTypePicker(actionType: actionType)
+            .equatable()
+
+        switch action {
+        case .arg0:
+            EmptyView()
+        case .arg1(.run):
+            ButtonMappingActionPickerRun(action: $action)
+        case .arg1(.mouseWheelScrollUp):
+            // TODO: TBD.
+            EmptyView()
+        case .arg1(.mouseWheelScrollDown):
+            // TODO: TBD.
+            EmptyView()
+        case .arg1(.mouseWheelScrollLeft):
+            // TODO: TBD.
+            EmptyView()
+        case .arg1(.mouseWheelScrollRight):
+            // TODO: TBD.
+            EmptyView()
+        case .arg1(.keyPress):
+            // TODO: TBD.
+            EmptyView()
+        }
+    }
+}
+
+extension ButtonMappingAction {
+    var actionType: Binding<ButtonMappingActionTypePicker.ActionType> {
+        Binding {
+            switch action {
+            case let .arg0(value):
+                return .arg0(value)
+            case .arg1(.run):
+                return .run
+            case .arg1(.mouseWheelScrollUp):
+                return .mouseWheelScrollUp
+            case .arg1(.mouseWheelScrollDown):
+                return .mouseWheelScrollDown
+            case .arg1(.mouseWheelScrollLeft):
+                return .mouseWheelScrollLeft
+            case .arg1(.mouseWheelScrollRight):
+                return .mouseWheelScrollRight
+            case .arg1(.keyPress):
+                return .keyPress
+            }
+        } set: { action in
+            switch action {
+            case let .arg0(value):
+                self.action = .arg0(value)
+            case .run:
+                self.action = .arg1(.run(""))
+            case .mouseWheelScrollUp:
+                self.action = .arg1(.mouseWheelScrollUp(.line(3)))
+            case .mouseWheelScrollDown:
+                self.action = .arg1(.mouseWheelScrollDown(.line(3)))
+            case .mouseWheelScrollLeft:
+                self.action = .arg1(.mouseWheelScrollLeft(.line(3)))
+            case .mouseWheelScrollRight:
+                self.action = .arg1(.mouseWheelScrollRight(.line(3)))
+            case .keyPress:
+                self.action = .arg1(.keyPress([]))
+            }
+        }
+    }
+}
